Create dependencies folder & helpers for tests + serverHost
