+++ package.json
@@ -45,8 +45,5 @@
   "workspaces": [
     "apps/*",
     "packages/**"
-  ],
-  "dependencies": {
-    "@asyncapi/modelina": "^5.2.5"
-  }
+  ]
 }
