+++ packages/helpers/src/index.js
@@ -1,17 +1,22 @@
 const { getMessageExamples, getOperationMessages } = require('./operations');
-const { getServerUrl, getServer }  = require('./servers');
+const { getServerUrl, getServer, getServerHost }  = require('./servers');
 const { getClientName, listFiles, getInfo, toSnakeCase, getTitle} = require('./utils');
 const { getQueryParams } = require('./bindings');
+const { cleanTestResultPaths, verifyDirectoryStructure, getDirElementsRecursive} = require('./testing');
 
 module.exports = {
   getServerUrl,
   getClientName,
   getServer,
+  getServerHost,
   listFiles,
   getQueryParams,
   getOperationMessages,
   getMessageExamples,
   getTitle,
   getInfo,
-  toSnakeCase
+  toSnakeCase,
+  cleanTestResultPaths,
+  verifyDirectoryStructure,
+  getDirElementsRecursive
 };
