+++ packages/helpers/src/servers.js
@@ -44,9 +44,26 @@ const getServer = (servers, serverName) => {
   return servers.get(serverName);
 };
 
+/**
+   * Get server Host from AsyncAPI server object.
+   *
+   * @param {object} server - The AsyncAPI server object.
+   * 
+   * return {string} - The server host.
+   */
+const getServerHost = (server) => {
+  let serverHost = server.host();
+  if (!serverHost) {
+    console.log('ERROR: host found in the server configuration.');
+    return null;
+  }
+  return serverHost;
+};
+  
 module.exports = {
   getServerUrl,
-  getServer
+  getServer,
+  getServerHost
 };
 
 //TODO: this separate file for helpers for servers represents approach to keep all helpers in separate files related to extractions of data from specific high level AsyncAPI objects. Here we will have more helpers for example related to variables extraction from servers, security, etc.
