+++ packages/templates/clients/websocket/java/quarkus/components/ClientConnector.js
@@ -2,7 +2,7 @@ import { Text } from '@asyncapi/generator-react-sdk';
 import InitConnector from './InitConnector';
 import { ConnectorFields } from './ConnectorFields';
 
-export default function ClientConnector({clientName}) {
+export default function ClientConnector({ clientName }) {
   return (
     <Text newLines={2} indent={2}>
       <Text newLines={2}>
