+++ packages/templates/clients/websocket/java/quarkus/components/ConnectorFields.js
@@ -1,6 +1,6 @@
 import { Text } from '@asyncapi/generator-react-sdk';
 
-export function ConnectorFields({ clientName}) {
+export function ConnectorFields({ clientName }) {
   return (
     <Text indent={2} newLines={2}>
       {`@Inject
