+++ packages/templates/clients/websocket/java/quarkus/components/dependencies/ConnectorDependencies.js
@@ -12,7 +12,6 @@ import io.quarkus.websockets.next.WebSocketClientConnection;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
 import jakarta.annotation.PostConstruct;
-import java.net.URI;
 import io.quarkus.logging.Log;
 import io.quarkus.runtime.Startup;`}
       </Text>
