+++ packages/templates/clients/websocket/java/quarkus/package.json
@@ -14,7 +14,6 @@
   "scripts": {
     "test": "jest  --coverage --passWithNoTests",
     "test:update": "npm run test -- -u",
-    "test:cleanup": "rimraf \"test/temp\"",
     "lint": "eslint --max-warnings 0 --config ../../../../../../.eslintrc --ignore-path ../../../../../../.eslintignore .",
     "lint:fix": "eslint --fix --max-warnings 0 --config ../../../../../../.eslintrc --ignore-path ../../../../../../.eslintignore ."
   },
@@ -34,8 +33,7 @@
     "eslint-plugin-jest": "^23.8.2",
     "eslint-plugin-react": "^7.34.1",
     "eslint-plugin-sonarjs": "^0.5.0",
-    "jest-esm-transformer": "^1.0.0",
-    "rimraf": "^3.0.2"
+    "jest-esm-transformer": "^1.0.0"
   },
   "jest": {
     "moduleFileExtensions": [
