+++ packages/templates/clients/websocket/java/quarkus/template/src/main/java/com/asyncapi/connector.java.js
@@ -1,11 +1,11 @@
 import { getClientName } from '@asyncapi/generator-helpers';
 import { File } from '@asyncapi/generator-react-sdk';
-import { ConnectorDependencies } from '../../../../../../components/ConnectorDependencies.js';
+import { ConnectorDependencies } from '../../../../../../components/dependencies/ConnectorDependencies.js';
 import ClientConnector from '../../../../../../components/ClientConnector.js';
 
 export default async function ({ asyncapi, params }) {
   const clientName = getClientName(asyncapi, params.appendClientSuffix, params.customClientName);
-  const clientConnectorName = `${clientName  }Connector.java`;
+  const clientConnectorName = `${clientName}Connector.java`;
   
   return (
     <File name={clientConnectorName}>
