+++ packages/templates/clients/websocket/java/quarkus/template/src/main/resources/AppProperties.js
@@ -1,14 +1,10 @@
-import { getClientName, getServer } from '@asyncapi/generator-helpers';
+import { getClientName, getServer, getServerHost} from '@asyncapi/generator-helpers';
 import { Text, File } from '@asyncapi/generator-react-sdk';
 
 export default function AppProperties({ asyncapi, params }) {
   const server = getServer(asyncapi.servers(), params.server);
   const clientName = getClientName(asyncapi, params.appendClientSuffix, params.customClientName);
-  const serverHost = server.host();
-  if (!serverHost) {
-    console.log('ERROR: host found in the server configuration.');
-    return null;
-  }
+  const serverHost = getServerHost(server);
 
   return (
     <File name="application.properties">
